test: add ramm getInternalPrice to test/utils and utilise it
