+++ lib/protocol.js
@@ -14,90 +14,6 @@ const COVER_PRICE_DENOMINATOR = 10000n; // 100% bps
 const REWARDS_DENOMINATOR = 10000n; // 100% bps
 const BUCKET_DURATION = BigInt(28 * 24 * 3600); // 28 days
 
-function getObservationIndex(timestamp, { PERIOD_SIZE, GRANULARITY }) {
-  if (timestamp % PERIOD_SIZE > 1) {
-    return (timestamp / PERIOD_SIZE + 1n) % GRANULARITY;
-  }
-  return (timestamp / PERIOD_SIZE) % GRANULARITY;
-}
-
-// ======================== RAMM ==============================================
-
-// TODO: eject from lib - we'll clearly not use it outside of tests
-/**
- * Calculates the internal NXM token price in ETH for given states
- */
-function calculateInternalPrice(currentState, observations, capital, supply, currentTimestamp, constants) {
-  const { GRANULARITY } = constants;
-  const currentIdx = getObservationIndex(BigInt(currentTimestamp), constants);
-  const previousIdx = (currentIdx + 1n) % GRANULARITY;
-
-  const firstObservation = observations[Number(previousIdx)];
-  const currentObservation = observations[Number(currentIdx)];
-
-  const elapsed = BigInt(currentTimestamp) - firstObservation.timestamp;
-
-  const spotPriceA = (parseEther('1') * currentState.eth) / currentState.nxmA;
-  const spotPriceB = (parseEther('1') * currentState.eth) / currentState.nxmB;
-
-  const averagePriceA = (currentObservation.priceCumulativeAbove - firstObservation.priceCumulativeAbove) / elapsed;
-
-  const averagePriceB = (currentObservation.priceCumulativeBelow - firstObservation.priceCumulativeBelow) / elapsed;
-
-  const priceA = BigIntMath.min(averagePriceA, spotPriceA);
-  const priceB = BigIntMath.max(averagePriceB, spotPriceB);
-  const bookValue = (parseEther('1') * capital) / supply;
-
-  const internalPrice = priceA + priceB - bookValue;
-  const maxPrice = (parseEther('1') * 3n * capital) / supply; // 300% BV
-  const minPrice = (parseEther('1') * 35n * capital) / (supply * 100n); // 35% BV
-
-  return BigIntMath.max(BigIntMath.min(internalPrice, maxPrice), minPrice);
-}
-
-// TODO: eject from lib - we'll clearly not use it outside of tests
-/**
- * Calculates the expected internal NXM price in ETH
- */
-async function getInternalPrice(ramm, pool, tokenController, timestamp) {
-  const capital = await pool.getPoolValueInEth();
-  const supply = await tokenController.totalSupply();
-  const mcrValue = await pool.getMCR();
-  const context = {
-    capital,
-    supply,
-    mcr: mcrValue,
-  };
-
-  const GRANULARITY = await ramm.GRANULARITY();
-  const PERIOD_SIZE = await ramm.PERIOD_SIZE();
-  const previousState = await ramm.loadState();
-  const previousObservations = [];
-
-  for (let i = 0; i < 3; i++) {
-    const observation = await ramm.observations(i);
-    previousObservations[i] = {
-      priceCumulativeAbove: observation.priceCumulativeAbove,
-      priceCumulativeBelow: observation.priceCumulativeBelow,
-      timestamp: observation.timestamp,
-    };
-  }
-
-  const state = {
-    nxmA: previousState.nxmA,
-    nxmB: previousState.nxmB,
-    eth: previousState.eth,
-    budget: previousState.budget,
-    ratchetSpeedB: previousState.ratchetSpeedB,
-    timestamp: previousState.timestamp,
-  };
-
-  const [currentState] = await ramm._getReserves(state, context, timestamp);
-  const observations = await ramm._updateTwap(state, previousObservations, context, timestamp);
-
-  return calculateInternalPrice(currentState, observations, capital, supply, timestamp, { GRANULARITY, PERIOD_SIZE });
-}
-
 // ======================= COVER ==============================================
 
 /**
@@ -283,8 +199,6 @@ async function stake({ contracts, stakingPool, staker, productId, period, graceP
 }
 
 module.exports = {
-  calculateInternalPrice,
-  getInternalPrice,
   calculatePremium,
   calculateRewards,
   calculateCoverEditPremium,
